Update README and finalize db_update script relocation
