+++ README.md
@@ -19,7 +19,7 @@ In a web browser, go to `http://127.0.0.1:8000`
 
 ## Adding Data to the Database
 * Configure database update script:  
-Open `config.txt` in a text editor and specify the path to the data under the variable `data_dir`
-* Run `db_update.py` script
+Open `config.json` in a text editor and specify the path to the data under the variable "data_dir"
+* Run `src/db_update.py` script
 * Check website _Directory_ tab for the new data:  
-If already on _Directory_, you will have to change to a different tab such as _Home_ and go back to _Directory_
+If already on _Directory_, you will have to refresh the page.
