+++ src/db_update.py
@@ -1,49 +1,102 @@
 """
-Updates the database using Sqlite to match the folder
-structure of the specified directory found in config.txt
+Updates the PostgreSQL database using psycopg2 to match the folder structure of the specified
+directory found in config.json
 """
 import os
-import json
-import sqlite3
+import re
+import argparse
 import subprocess
+from time import time
+from io import StringIO
 from threading import Lock
+from typing import Callable
 from functools import partial
 from concurrent.futures import ThreadPoolExecutor
 
+import psycopg2
 import numpy as np
+import pandas as pd
 from numpy import ndarray
+from psycopg2.extras import execute_values
 
 from src.utils.utils import progress_bar
+from nicer_website.settings import DATABASES, DATA_DIR
 
 
-def table_insert(data: dict[str, list[int | float | str | None]], batch_size: int = 100) -> None:
+DB_CONFIG: dict[str, str] = {
+    'host': DATABASES['default'].get('HOST', 'localhost'),
+    'port': DATABASES['default'].get('PORT', '5432'),
+    'dbname': DATABASES['default'].get('NAME', ''),
+    'user': DATABASES['default'].get('USER', ''),
+    'password': DATABASES['default'].get('PASSWORD', ''),
+}
+
+
+def create_table(data: dict[str, list[str | None]], batch_size: int = 10000) -> None:
     """
     Add folder and file data to the database, including additional metadata fields.
 
     Parameters
     ----------
-    data : dict[str, list[int | float | str | None]]
+    data : dict[str, list[str | None]]
         Dictionary containing the data to be inserted into the database, keys are column names and
-        values are lists of corresponding values.
+        values are lists of corresponding values
+    batch_size : int, default = 10000
+        How many entries to insert into the database per execution
+    """
+    i: int
+    df = pd.DataFrame.from_dict(data)
+    df.fillna('null', inplace=True)
+
+    # Connect to the PostgreSQL database
+    with psycopg2.connect(**DB_CONFIG) as conn:
+        with conn.cursor() as cur:
+            for i in range(0, len(df), batch_size):
+                batch = df.iloc[i:i + batch_size]
+                sio = StringIO()
+                batch.to_csv(sio, index=False, header=False)
+                sio.seek(0)
+
+                cur.copy_expert(
+                    f"COPY file_mgr_item ({', '.join(df)}) FROM STDIN WITH "
+                    f"(FORMAT CSV, NULL 'null');",
+                    sio,
+                )
+                progress_bar(i + len(batch), len(df))
+
+
+def table_insert(data: dict[str, list[str | None]], batch_size: int = 10000) -> None:
+    """
+    Add folder and file data to the database, including additional metadata fields.
+
+    Parameters
+    ----------
+    data : dict[str, list[str | None]]
+        Dictionary containing the data to be inserted into the database, keys are column names and
+        values are lists of corresponding values
     batch_size : int, default = 100
         How many entries to insert into the database per execution
     """
     i: int
-    update: str = f'INSERT OR REPLACE INTO file_mgr_item ({", ".join(data)})' \
-                  f'VALUES ({", ".join(["?"] * len(data))})'
+    update: str = f'INSERT INTO file_mgr_item ({", ".join(data)}) ' \
+                  f'VALUES %s ON CONFLICT (name, path, type) DO UPDATE SET ' + \
+                  ', '.join(
+                      [f'{col}=EXCLUDED.{col}' for col in data
+                       if col not in ['name', 'path', 'type']],
+                  )
     batch: ndarray[tuple[int, int], np.dtype[np.object_]]
     batches: list[ndarray[tuple[int, int], np.dtype[np.object_]]] = np.array_split(
         np.array(list(data.values())),
-        int(len(list(data.values())[0]) / batch_size),
+        max(int(len(list(data.values())[0]) / batch_size), 1),
         axis=-1,
     )
 
-    # Connect to the database
-    with sqlite3.connect('db.sqlite3') as conn:
-        # Insert data into the database
-        for i, batch in enumerate(batches):
-            conn.executemany(update, batch.swapaxes(0, 1))
-            progress_bar(i, len(batches))
+    # Connect to the PostgreSQL database
+    with psycopg2.connect(**DB_CONFIG) as conn:
+        with conn.cursor() as cur:
+            for i, batch in enumerate(batches):
+                execute_values(cur, update, batch.swapaxes(0, 1), page_size=batch.shape[-1])
+                progress_bar(i, len(batches))
 
 
 def linux_count(directory: str) -> int:
@@ -98,18 +151,15 @@ def universal_count(directory: str) -> int:
 
 
 def process_dir(
-        data_dir: str,
         count: list[int],
         total: list[int],
         root_file: tuple[str, ndarray[tuple[int], np.dtype[np.str_]]],
-        lock: Lock) -> dict[str, list[int | float | str | None]]:
+        lock: Lock) -> dict[str, list[str | None]]:
     """
     Process a directory and its files to extract metadata and prepare for database insertion.
 
     Parameters
     ----------
-    data_dir : str
-        The base directory for data.
     count : list[int]
         A list containing the current count of processed items, used for thread-safe updates.
     total : list[int]
@@ -122,22 +172,19 @@ def process_dir(
 
     Returns
     -------
-    dict[str, list[int | float | str]]
-        A dictionary containing metadata extracted from the directory and files.
+    pd.DataFrame
+        A DataFrame containing metadata extracted from the directory and files.
     """
     dets: int = 52
     min_num: int = 1000
     file: str
     root: str = root_file[0]
-    relative_root: str = root.replace(data_dir, '')
-    dir_name: str = os.path.basename(relative_root)
-    parent_dir: str = os.path.dirname(relative_root) + '/'
-    relative_root += '/'
-    data: dict[str, list[int | float | str | None]] = {key: [] for key in [
-        'name',
-        'path',
-        'type',
-        'source',
+    relative_root: str = root.replace(DATA_DIR, '').lstrip('/') or '/'
+    dir_name: str = os.path.basename(relative_root) or '/'
+    parent_dir: str = os.path.dirname(relative_root) or '/'
+    keys: tuple[str, ...] = (
+        'object',
+        'obsid',
         'tstart_tt',
         'tstop_tt',
         'ra',
@@ -145,80 +192,70 @@ def process_dir(
         'ndets_used',
         'ushoot_net_rate',
         'oshoot_net_rate',
-        'changegoodx_5_12_rate',
-    ]}
-    summary: ndarray[tuple[int], np.dtype[np.str_]]
+        'goodx_0p5_12_rate',
+    )
+    key_map: dict[str, str] = {'object': 'source', 'obsid': 'obs_id'}
+    data: dict[str, list[str | None]]
+    summary: dict[str, list[str | None]]
+    gti: re.Match[str] | None
+    mask: ndarray[tuple[int], np.dtype[np.bool_]]
+    quality: ndarray[tuple[int], np.dtype[np.str_]]
     files: ndarray[tuple[int], np.dtype[np.str_]] = root_file[1]
     summaries: ndarray[tuple[int], np.dtype[np.str_]] = files[np.char.endswith(files, '.summary')]
+    qualities: ndarray[tuple[int], np.dtype[np.str_]] = np.array(
+        ['goddard', 'gold', 'silver', 'radium', 'pyrite'],
+    )
 
     # Remove ARF and RMF files from the list of files and decrement the total count as these
     # aren't needed for now
     if len(files):
+        mask = (np.char.find(files, '.arf') != -1) | (np.char.find(files, '.rmf') != -1)
+
         with lock:
-            total[0] -= np.count_nonzero(np.char.find(files, '.arf') != -1)
-            total[0] -= np.count_nonzero(np.char.find(files, '.rmf') != -1)
+            total[0] -= np.count_nonzero(mask)
 
-        files = np.delete(np.array(files), np.char.find(files, '.arf') != -1)
-        files = np.delete(np.array(files), np.char.find(files, '.rmf') != -1)
+        files = np.delete(files, mask)
 
-    # Find source name if in a destination that contains source information
     if len(summaries):
-        summary = np.loadtxt(f'{root}/{summaries[0]}', dtype=str)
-        data['source'].extend([
-            summary[:, 1][summary[:, 0] == 'OBJECT'][0].strip("'")
-        ] * (len(files) + bool(dir_name)))
-        data['tstart_tt'].extend([
-            float(summary[:, 1][summary[:, 0] == 'TSTART_TT'][0].strip())
-        ] * (len(files) + bool(dir_name)))
-        data['tstop_tt'].extend([
-            float(summary[:, 1][summary[:, 0] == 'TSTOP_TT'][0].strip())
-        ] * (len(files) + bool(dir_name)))
-        data['ra'].extend([
-            float(summary[:, 1][summary[:, 0] == 'RA'][0].strip())
-        ] * (len(files) + bool(dir_name)))
-        data['dec'].extend([
-            float(summary[:, 1][summary[:, 0] == 'DEC'][0].strip())
-        ] * (len(files) + bool(dir_name)))
-        data['ndets_used'].extend([
-            float(summary[:, 1][summary[:, 0] == 'NDETS_USED'][0].strip())
-        ] * (len(files) + bool(dir_name)))
-        data['ushoot_net_rate'].extend([
-            float(summary[:, 1][summary[:, 0] == 'USHOOT_NET_RATE'][0].strip()),
-        ] * (len(files) + bool(dir_name)))
-        data['oshoot_net_rate'].extend([
-            float(summary[:, 1][summary[:, 0] == 'OSHOOT_NET_RATE'][0].strip()),
-        ] * (len(files) + bool(dir_name)))
-        data['changegoodx_5_12_rate'].extend([
-            float(summary[:, 1][summary[:, 0] == 'GOODX_0p5_12_RATE'][0].strip()) * dets,
-        ] * (len(files) + bool(dir_name)))
+        summary = {key.lower(): [val.strip("'")] for key, val in np.loadtxt(
+            os.path.join(root, summaries[0]),
+            dtype=str,
+        ) if key.lower() in keys}
+
+        if summary['goodx_0p5_12_rate'][0]:
+            summary['goodx_0p5_12_rate'][0] = str(float(summary['goodx_0p5_12_rate'][0]) * dets)
     else:
-        # Default values for additional metadata if not available
-        data['source'].extend([''] * (len(files) + bool(dir_name)))
-        data['tstart_tt'].extend([None] * (len(files) + bool(dir_name)))
-        data['tstop_tt'].extend([None] * (len(files) + bool(dir_name)))
-        data['ra'].extend([None] * (len(files) + bool(dir_name)))
-        data['dec'].extend([None] * (len(files) + bool(dir_name)))
-        data['ndets_used'].extend([None] * (len(files) + bool(dir_name)))
-        data['ushoot_net_rate'].extend([None] * (len(files) + bool(dir_name)))
-        data['oshoot_net_rate'].extend([None] * (len(files) + bool(dir_name)))
-        data['changegoodx_5_12_rate'].extend([None] * (len(files) + bool(dir_name)))
+        summary = {key: ['' if key == 'object' else None] for key in keys}
 
-    # If not top-level directory, add folder to the database
-    if dir_name:
-        data['name'].append(dir_name)
-        data['path'].append(parent_dir)
-        data['type'].append('dir')
+    for key, mapped_key in key_map.items():
+        if key in summary:
+            summary[mapped_key] = summary.pop(key)
 
-        with lock:
-            count[0] += 1
+    # Load data
+    data = {
+        'gti': [None],
+        'name': [dir_name],
+        'quality': [''],
+        'path': [parent_dir],
+        'file_type': [''],
+        'type': ['dir'],
+    }
 
-            if count[0] % min_num == 0:
-                progress_bar(count[0], total[0])
+    with lock:
+        count[0] += 1
+
+        if count[0] % min_num == 0:
+            progress_bar(count[0], total[0])
 
     # Add file to the database
     for file in files:
+        data['gti'].append(gti.group(1) if (gti := re.search(r'GTI(\d+)', file)) else None)
         data['name'].append(file)
+        data['quality'].append(
+            quality[0] if len(quality := qualities[np.char.find(file, qualities) != -1]) else '',
+        )
         data['path'].append(relative_root)
+        data['file_type'].append(file.split('.', 1)[-1] if '.' in file else '')
         data['type'].append('file')
 
         with lock:
@@ -226,59 +263,111 @@ def process_dir(
 
             if count[0] % min_num == 0:
                 progress_bar(count[0], total[0])
+
+    for key, value in summary.items():
+        data[key] = value * len(data['name'])
     return data
 
 
-def main() -> None:
+def main(update: bool = False, batch: int = int(1e5), limit: int = -1) -> None:
     """
     Main function for updating the database
+
+    Parameters
+    ----------
+    update : bool, default = False
+        Whether to update existing entries in the database
+    batch : int, default = 100000
+        Batch size for database insertion
+    limit : int, default = -1
+        Limit the number of files/folders processed, -1 for no limit
     """
     min_num: int = 1000
-    key: str
-    data_dir: str
+    ti: float = time()
     count: list[int] = [0]
     total: list[int] = [0]
-    value: list[int | float | str | None]
     root_files: list[tuple[str, ndarray[tuple[int], np.dtype[np.str_]]]] = []
-    results: list[dict[str, list[int | float | str | None]]]
-    data: dict[str, list[int | float | str | None]]
-    result: dict[str, list[int | float | str | None]]
+    results: list[dict[str, list[str | None]]]
+    data: dict[str, list[str | None]]
+    func: Callable[[dict[str, list[str | None]], int], None]
     lock: Lock = Lock()
-    partial_func: partial[dict[str, list[int | float | str | None]]]
-    #os.chdir('../')
+    partial_func: partial[dict[str, list[str | None]]]
 
-    # Get data directory location from config.txt
-    with open('config.txt', mode='r', encoding='utf-8') as config:
-        data_dir = json.load(config)['data_dir']
+    if not update:
+        with psycopg2.connect(**DB_CONFIG) as conn:
+            with conn.cursor() as cur:
+                cur.execute("SELECT EXISTS (SELECT 1 FROM file_mgr_item LIMIT 1);")
+
+                if cur.fetchone()[0]:
+                    raise ValueError('Database already populated, use update=True to update '
+                                     'existing entries or clear database with: python manage.py '
+                                     'flush')
 
     # Calculate the total number of folders and files
-    for root, _, files in os.walk(data_dir):
+    for root, _, files in os.walk(DATA_DIR):
         root_files.append((root, np.array(files, dtype=str)))
-        total[0] += len(files) + int(root != data_dir)
+        total[0] += len(files) + int(root != DATA_DIR)
 
         if total[0] % min_num == 0:
             print(f'\rCount: {total[0]}', end='', flush=True)
 
-    if not total:
-        raise ValueError(f'No files or folders found, check parent directory is correct: '
-                         f'{data_dir}')
+        if limit > 0 and total[0] >= limit:
+            print(f'\nLimit of {limit} reached, stopping count early...')
+            break
 
-    print(f'\rTotal number of files and folders: {total[0]}')
+    if not total[0]:
+        raise ValueError(f'No files or folders found, check parent directory is correct: '
+                         f'{DATA_DIR}')
+
+    print(f'\rTotal number of files and folders: {total[0]}\tTime taken: {time() - ti:.2f} s')
+    ti = time()
+    partial_func = partial(process_dir, count, total, lock=lock)
+    print('\nProcessing files and folders...')
 
     with ThreadPoolExecutor() as executor:
-        partial_func = partial(process_dir, data_dir, count, total, lock=lock)
         results = list(executor.map(partial_func, root_files))
 
-    progress_bar(count[0], total[0])
-    data = {key: [] for key in results[0].keys()}
+    progress_bar(count[0], total[0], text=f'Time taken: {time() - ti:.2f} s')
+    ti = time()
+    print('\nCombining results...')
+    data = {}
 
     for result in results:
-        for key, value in result.items():
-            data[key].extend(value)
+        if not data:
+            for key, val in result.items():
+                data[key] = val
+        else:
+            for key, val in result.items():
+                data[key].extend(val)
+
+    print(f'\nData combining complete, time taken: {time() - ti:.2f} s')
 
     # Insert data into database
-    table_insert(data)
+    ti = time()
+    print('\nInserting data into database...')
+    func = table_insert if update else create_table
+    func(data, batch_size=batch)
+    print(f'\nDatabase update complete\tTime taken: {time() - ti:.2f} s')
 
 
 if __name__ == '__main__':
-    main()
+    parser = argparse.ArgumentParser(description="Update the file_mgr database.")
+    parser.add_argument(
+        '--update',
+        action='store_true',
+        help='Whether to update existing entries in the database',
+    )
+    parser.add_argument(
+        '--batch-size',
+        type=int,
+        default=int(1e5),
+        help='Batch size for database insertion',
+    )
+    parser.add_argument(
+        '--limit',
+        type=int,
+        default=-1,
+        help='Limit the number of files/folders processed, -1 for no limit',
+    )
+    args = parser.parse_args()
+    main(update=args.update, batch=args.batch_size, limit=args.limit)
